+++ processor/src/test/fixtures/bad_input/com/example/injector/outputs/TypedOutputModel.java
@@ -1,0 +1,17 @@
+package com.example.injector.outputs;
+
+import sting.Injectable;
+import sting.Injector;
+import sting.Typed;
+
+@Injector
+interface TypedOutputModel
+{
+  @Typed( MyModel.class )
+  MyModel getMyModel();
+
+  @Injectable
+  class MyModel
+  {
+  }
+}