+++ processor/src/test/fixtures/input/com/example/meta/ActAsStingComponentOnlyModel.java
@@ -1,0 +1,15 @@
+package com.example.meta;
+
+import sting.ActAsStingComponent;
+
+public final class ActAsStingComponentOnlyModel
+{
+  @ActAsStingComponent
+  @interface FrameworkComponent
+  {
+  }
+
+  private ActAsStingComponentOnlyModel()
+  {
+  }
+}